+++ examples/bss_example.py
@@ -32,6 +32,13 @@ The method implemented is described in the following publication
     K. Sekiguchi, A. A. Nugraha, Y. Bando, K. Yoshii, *Fast Multichannel Source 
     Separation Based on Jointly Diagonalizable Spatial Covariance Matrices*, EUSIPCO, 2019.
 
+5) Fast Multichannel Nonnegative Matrix Factorization 2 (FastMNMF2)
+The method implemented is described in the following publication
+
+    K. Sekiguchi, Y. Bando, A. A. Nugraha, K. Yoshii, T. Kawahara, *Fast Multichannel Nonnegative
+    Matrix Factorization With Directivity-Aware Jointly-Diagonalizable Spatial
+    Covariance Matrices for Blind Source Separation*, IEEE/ACM TASLP, 2020.
+
 All the algorithms work in the STFT domain. The test files were extracted from the
 `CMU ARCTIC <http://www.festvox.org/cmu_arctic/>`_ corpus.
 
@@ -69,13 +76,13 @@ wav_files = [
 
 if __name__ == "__main__":
 
-    choices = ["ilrma", "auxiva", "sparseauxiva", "fastmnmf"]
+    choices = ["ilrma", "auxiva", "sparseauxiva", "fastmnmf", "fastmnmf2"]
 
     import argparse
 
     parser = argparse.ArgumentParser(
         description="Demonstration of blind source separation using "
-        "IVA, ILRMA, or sparse IVA ."
+        "IVA, ILRMA, sparse IVA, FastMNMF, or FastMNMF2 ."
     )
     parser.add_argument("-b", "--block", type=int, default=2048, help="STFT block size")
     parser.add_argument(
@@ -192,7 +199,12 @@ if __name__ == "__main__":
     elif bss_type == "fastmnmf":
         # Run FastMNMF
         Y = pra.bss.fastmnmf(
-            X, n_iter=30, n_components=8, n_src=2, callback=convergence_callback
+            X, n_iter=30, n_components=2, n_src=2, callback=convergence_callback
+        )
+    elif bss_type == "fastmnmf2":
+        # Run FastMNMF2
+        Y = pra.bss.fastmnmf2(
+            X, n_iter=30, n_components=2, n_src=2, callback=convergence_callback
         )
     elif bss_type == "sparseauxiva":
         # Estimate set of active frequency bins